+++ src/gui/Coordinator.hpp
@@ -38,6 +38,11 @@ namespace mdl {
 
     void update_current_filter_if_necessary();
 
+    void on_previous_filter();
+    void on_next_filter();
+
+    void set_scroll_filter(bool state);
+
   private:
     FilterList& filter_list_;
     Glib::RefPtr<FilterListModel> filter_model_;
@@ -49,6 +54,7 @@ namespace mdl {
     FilterPanelFactory panel_factory_;
     FilterPanel* current_filter_panel_;
     fg::Filter* current_filter_;
+    bool scroll_filter_;
 
     sigc::connection on_filter_selected_;
     void on_filter_selected(int start_frame);